zbuild: separate helper functions to separate file
